% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_nobs.R
\name{add_nobs}
\alias{add_nobs}
\title{add_nobs
add number of observed subtrips according to the `DISCARD_SOURCE` used
`n` is number of obsereved subtrips used. This may cross fishing years in the case of Assumed (A) or transition rates (T). Gear only rates use two years of information so `n` for those is across two years (\year_{t} and \year_{t-1})}
\usage{
add_nobs(joined_table)
}
\arguments{
\item{joined_table}{}
}
\value{
dataframe from input with additional columns for n_USED and N_USED
}
\description{
add_nobs
add number of observed subtrips according to the `DISCARD_SOURCE` used
`n` is number of obsereved subtrips used. This may cross fishing years in the case of Assumed (A) or transition rates (T). Gear only rates use two years of information so `n` for those is across two years (\year_{t} and \year_{t-1})
}
\author{
Benjamin Galuardi
}
