% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_discard_all_years.R
\name{append_discard_all_years}
\alias{append_discard_all_years}
\title{append_discard_all_years: Delete and fill}
\usage{
append_discard_all_years(con = con, data = outlist, drop = FALSE, gf_only = F)
}
\arguments{
\item{con}{ROracle connection to Oracle (e.g. MAPS)}

\item{data}{Dataframe to delete and insert into Oracle discard_all_years table}

\item{drop}{Logical whether to drop and recreate the cams_discard_all_years table}
}
\description{
append_discard_all_years: Delete and fill
}
