% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_discard_source.R
\name{assign_discard_source}
\alias{assign_discard_source}
\title{Assign Discard Source
this is an internal function that assigns DISCARD_SOURCE. IT is explcitly designed to account for situations where there is an observed trip with no observed hauls.}
\usage{
assign_discard_source(jtable, GF = 1)
}
\arguments{
\item{jtable}{joined table created during the discard run}

\item{GF}{0/1 denoting groundfish trips or not. Groundfish trips have a different set of DISCARD_SOURCE}
}
\value{
data frame (joined_table in discard run)
}
\description{
Assign Discard Source
this is an internal function that assigns DISCARD_SOURCE. IT is explcitly designed to account for situations where there is an observed trip with no observed hauls.
}
