% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_strata.R
\name{assign_strata}
\alias{assign_strata}
\title{Assign Strata
uses text input to coalesce variables to a STRATA column}
\usage{
assign_strata(dat, stratvars)
}
\arguments{
\item{dat}{catch or observation input}

\item{stratvars}{variables in `dat` to coalesce}
}
\value{
a data frame (dat) with a `STRATA` column
}
\description{
Assign Strata
uses text input to coalesce variables to a STRATA column
}
