% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootr.strat.r
\name{bootr.strat}
\alias{bootr.strat}
\title{Bootstrap discard rate with stratification
bootstraps Cochran discard rate from a set of observed trips}
\usage{
bootr.strat(bdat, ddat, focal_year = 2015, strata_name = "strata",
  strata_complete = NULL, coverage = NULL, keep.samps = F, ...)
}
\arguments{
\item{bdat}{observer data with strata column}

\item{ddat}{dmis data (commercial trips) with strata column}

\item{coverage}{proportion of commercial trips observed (i.e., observer coverage), which can optionally be specified to explore different coverages [default is observed coverage]}

\item{keep.samps}{option to keep samples for each bootstrap (default = F)}

\item{...}{additional arguments to \code{\link{cochran.trans.calc}} (i.e. strata_complete and strata_name, interval)}

\item{days}{number of days to calculate starting from the first day of the fishing year}
}
\value{
list of resampled discard rates
}
\description{
Bootstrap discard rate with stratification
bootstraps Cochran discard rate from a set of observed trips
}
\details{
This function runs a single instance with resampled observer data. It may be parallelized easily (see example). This requires mcoptions from \code{\link{setup.parallel}}. Including strata_complete is strongly recemmended.
}
\examples{
nboot = 10
mcoptions = setup.parallel()
bout.list = foreach(1:nboot, .options.multicore = mcoptions) \%dopar\% {
library(discaRd)
discaRd::bootr.strat(bdat = bdat, ddat = ddat, focal_year = 2015, nboot = nboot, strata_name = 'strata', strata_complete = strata_complete)
}
plot(colSums(bout.list[[1]]), typ='l', ylim = c(0,100000),  xlab='Day of year', ylab = 'Discard')
for(i in 2:length(bout.list)){
lines(colSums(bout.list[[i]]), typ='l', col =2, lty = 2)
}


}
\seealso{
\code{\link{samp.by.strat}}
}

