% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cochran_rate.r
\name{calc_cochran_rate}
\alias{calc_cochran_rate}
\title{Estimates Cochran discard rate and required sample sizes}
\usage{
calc_cochran_rate(df, n_trips, n_obs, CV_targ = NA)
}
\arguments{
\item{df}{dataframe where each row is a trip. Requires LINK1, BYCATCH, KALL naming convention}

\item{n_trips}{Number of total trips}

\item{CV_targ}{target CV}
}
\value{
observed sample size needed for target CV
}
\description{
This function estimates the discard rate using the Cochran ratio estimator and returns the mean rate as well as estimates of uncertainty and required sample sizes for a target CV
}
\examples{

test_df <- data.frame(
 LINK1 = c("Trip_A", "Trip_B", "Trip_C", "Trip_D", "Trip_E"),
 BYCATCH = c(10, 50, 20, 10, 0),    # Discard weight
 KALL = c(100, 500, 150, 100, 200)  # Kept weight
)
# --- Step 2: Run the function ---
# Scenario: Total fleet size is 1000 trips, we observed 5. Target CV is 30\% (0.30)
result <- calc_cochran_rate(
 df = test_df,
 n_trips = 1000,
 n_obs = nrow(test_df),
 CV_targ = 0.30
)
# --- Step 3: View Results ---
print(result)
}
\references{
Wiley: Sampling Techniques, 3rd Edition - William G. Cochran. (n.d.). Available from http://www.wiley.com/WileyCDA/WileyTitle/productCd-047116240X.html [accessed 1 August 2016].

Wigley SE, Rago PJ, Sosebee KA, Palka DL. 2007. The Analytic Component to the Standardized Bycatch Reporting Methodology Omnibus Amendment: Sampling Design, and Estimation of Precision and Accuracy (2nd Edition). US Dep. Commer., Northeast Fish. Sci. Cent. Ref. Doc. 07-09; 156 p
}
