% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_discard.R
\name{clean_discard_fy}
\alias{clean_discard_fy}
\title{Parse and Upload Discard Table
Parses discard results, cleans the table (e.g. infinite) and uploads to Oracle. This is designed to be used for a series of single species (itis), for a single fishing year, that reside in a folder. This function will upload all species within the folder that correspond to the provided fishing year.}
\usage{
clean_discard_fy(filepath = getOption("maps.discardsPath"), FY = 2018)
}
\arguments{
\item{filepath}{path to .fst discard results}

\item{FY}{Fishing Year to upload, should correspond to those results in filepath}

\item{con}{ROracle connection to Oracle (e.g. MAPS)}
}
\value{
nothing; uploads a table to the MAPS schema
}
\description{
N.B.! This function will remove the previous table(s) on MAPS.
}
