% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cochran.calc.ss.r
\name{cochran.calc.ss}
\alias{cochran.calc.ss}
\title{Estimates Cochran discard rate and required sample sizes}
\usage{
cochran.calc.ss(l_df, l_N, l_CVtarg = NA)
}
\arguments{
\item{l_df}{dataframe where each row is a trip. Requires LINK1, BYCATCH, KALL naming convention}

\item{l_N}{Number of total trips}

\item{l_CVtarg}{target CV}
}
\value{
observed sample size needed for target CV
}
\description{
This function estimates the discard rate using the Cochran ratio estimator and returns the mean rate as well as estimates of uncertainty and required sample sizes for a target CV
}
\examples{
cochran.calc.ss
}
\references{
Wiley: Sampling Techniques, 3rd Edition - William G. Cochran. (n.d.). Available from http://www.wiley.com/WileyCDA/WileyTitle/productCd-047116240X.html [accessed 1 August 2016].

Wigley SE, Rago PJ, Sosebee KA, Palka DL. 2007. The Analytic Component to the Standardized Bycatch Reporting Methodology Omnibus Amendment: Sampling Design, and Estimation of Precision and Accuracy (2nd Edition). US Dep. Commer., Northeast Fish. Sci. Cent. Ref. Doc. 07-09; 156 p
}

