% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cochran.trans.calc.r
\name{cochran.trans.calc}
\alias{cochran.trans.calc}
\title{Estimate the in-season cumulative discard with a transition rate}
\usage{
cochran.trans.calc(
  bydat_focal,
  trips_focal,
  bydat_prev,
  trips_prev,
  strata_name = "STRATA",
  strata_complete = NULL,
  time_span = c(1, 365),
  time_inter = 1,
  CV_target = 0.3,
  trans_method = c("ntrips", "ntripsCV", "moving", "none")[1],
  trans_num = 5,
  trans_numCV = FALSE
)
}
\arguments{
\item{bydat_focal}{Observed trips for focal year (t)}

\item{trips_focal}{Trips from DMIS for focal year (t)}

\item{bydat_prev}{Observed trips for previous year (t-1)}

\item{trips_prev}{Trips from DMIS for previous year (t-1)}

\item{strata_name}{Name of the stratification column (default = "STRATA")}

\item{strata_complete}{Vector of complete names for all strata of interest}

\item{time_span}{Start and end days of the fishing season}

\item{time_inter}{The length of the time interval for cumulative calculation (default = 1 day)}

\item{trans_method}{Method for incorporating the transition rate, 
either using a minimum number of trips specified [default] or determined by previous year's CV, a moving window, or no transition.}

\item{trans_num}{Number used by the \code{trans_method} for determining transition, 
either number of trips (\code{trans_method='ntrips'})}

\item{trans_numCV}{Option to output the number of days in transition for}
}
\value{
A list of Cochran estimates by strata, Discards, Total Discard in the fishery (adjusted by transition rate)
}
\description{
Apply transition rate to adjust in-season discard rate estimates and calculate the cumulative discard for a set of trips and observed trips, according to a strata definition
}
\details{
Each data frame input MUST have matching stratification columns. This variable may be numeric, logical, factor or character.
 The cumulative rate can be calculated without any transition rate by using method = "none".
}
\examples{
cochran.trans.calc
}
\seealso{
\code{\link{get.cochran.ss.by.strat}}
}
