% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_discard_all_years.R
\name{create_discard_all_years}
\alias{create_discard_all_years}
\title{create_discard_all_years}
\usage{
create_discard_all_years(con, drop = FALSE)
}
\arguments{
\item{con}{Oracle connection (ROracle preferred)}

\item{path}{Path to discard fst files}
}
\value{
Oracle table of merged CAMS landings and observer data
}
\description{
create_discard_all_years
}
\examples{

\dontrun{
# unlock keyring

keyring::keyring_unlock("apsd_ma")
con_maps = apsdFuns::roracle_login(key_name = 'apsd_ma', key_service = 'maps')
create_discard_all_years(con = con_maps)
}
}
