% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discard.R
\name{discard}
\alias{discard}
\title{discard}
\usage{
discard(.data, ...)
}
\arguments{
\item{.data}{a tbl}

\item{...}{Logical predicates defined in terms of the variables in .data.
Multiple conditions are combined with &.
Only rows where the condition evaluates to TRUE are discarded.}
}
\value{
a tbl
}
\description{
discard
}
\examples{
library(dplyr)
df <- tibble(
x1 = sample(c(-1, 0, 1), size = 1000, replace = TRUE),
x2 = sample(c(-1, 0, 1), size = 1000, replace = TRUE),
x3 = sample(c(-1, 0, 1), size = 1000, replace = TRUE),
x4 = sample(c(-1, 0, 1), size = 1000, replace = TRUE))
discard(df, x1 == 0, x2 == 0, x3 == 0, x4 == 0)
}
