% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discard.R
\name{discard_all}
\alias{discard_all}
\alias{discard_at}
\alias{discard_if}
\title{discard all}
\usage{
discard_all(.tbl, .vars_predicate)

discard_at(.tbl, .vars, .vars_predicate)

discard_if(.tbl, .predicate, .vars_predicate)
}
\arguments{
\item{.tbl}{A tbl object}

\item{.vars_predicate}{See [dplyr::filter_all()]}

\item{.vars}{See [dplyr::filter_all()]}

\item{.predicate}{See [dplyr::filter_all()]}
}
\description{
discard all
}
\examples{

discard_all(mtcars, all_vars(. > 150))
discard_all(mtcars, any_vars(. > 150))

discard_at(mtcars, vars(starts_with("d")), any_vars((. \%\% 2) == 0))

}
