% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discard_calendar.R
\name{discard_calendar}
\alias{discard_calendar}
\title{discard_calendar: Calculate discards for January fishing year species}
\usage{
discard_calendar(
  con,
  species = species,
  FY = fy,
  non_gf_dat = non_gf_dat,
  save_dir = file.path(getOption("maps.discardsPath"), "calendar")
)
}
\arguments{
\item{con}{ROracle connection to Oracle (e.g. MAPS)}

\item{species}{dataframe with species info}

\item{FY}{Fishing Year}

\item{non_gf_dat}{Data frame of non-groundfish trips built from CAMS_OBS_CATCH and control script routine}

\item{save_dir}{Directory to save (and load saved) results}
}
\value{
nothing currently, writes out to fst files (add oracle?)
}
\description{
discard_calendar: Calculate discards for January fishing year species
}
