% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discard_generic.R
\name{discard_generic}
\alias{discard_generic}
\title{discard_generic: Calculate discards for January fishing year species}
\usage{
discard_generic(
  con = con_maps,
  species = species,
  FY = fy,
  all_dat = all_dat,
  save_dir = file.path(getOption("maps.discardsPath"), "calendar"),
  run_parallel = FALSE
)
}
\arguments{
\item{con}{ROracle connection to Oracle (e.g. MAPS)}

\item{species}{dataframe with species info}

\item{FY}{Fishing Year}

\item{all_dat}{Data frame of trips built from CAMS_OBS_CATCH and control script routine}

\item{save_dir}{Directory to save (and load saved) results}

\item{run_parallel}{option to run species discard calculations in parallel}
}
\value{
nothing currently, writes out to fst files (add oracle?)
}
\description{
discard_generic: Calculate discards for January fishing year species
}
