% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discard_generic_nbr.R
\name{discard_generic_nbr_diagnostic}
\alias{discard_generic_nbr_diagnostic}
\title{discard_generic_nbr_diagnostic: Calculate discards for January fishing year
Support tables in this version are function variables. This allows them to be modified and run for diagnostic testing. This function can output a summary of discard information by strata, including number of trips used, variance, CV, discard amount, and KALL. It can also output a table, by subtrip, with all information. This table can be summarized afterward as well. If testing several scenarios, it is recommended to only output the summary rather than the full subtrip table as each output is 1-2GB.}
\usage{
discard_generic_nbr_diagnostic(
  con = con_maps,
  species = species,
  FY = fy,
  all_dat = all_dat,
  return_table = TRUE,
  return_summary = FALSE,
  CAMS_GEAR_STRATA = CAMS_GEAR_STRATA,
  STOCK_AREAS = STOCK_AREAS,
  CAMS_DISCARD_MORTALITY_STOCK = CAMS_DISCARD_MORTALITY_STOCK,
  run_parallel = TRUE
)
}
\arguments{
\item{con}{Oracle connection}

\item{species}{data frame of species for evaluation. Can be a dataframe of multiple species or single row (preferred for testing)}

\item{FY}{fishing year for evaluation}

\item{all_dat}{input data sourced from CAMS_OBS_CATCH}

\item{return_table}{logical, should a table (very large!) of trip level info be returned?}

\item{return_summary}{logical, should a summary (very not large) be returned?}

\item{CAMS_GEAR_STRATA}{support table sourced from Oracle}

\item{STOCK_AREAS}{support table sourced from Oracle}

\item{CAMS_DISCARD_MORTALITY_STOCK}{support table sourced from Oracle}
}
\description{
The original function pushed .fst files directly to the Wind server at GARFO. This version does not do that and only produces local results.
}
\author{
Benjamin Galuardi
}
