% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discard_groundfish.R
\name{discard_groundfish}
\alias{discard_groundfish}
\title{discard_groundfish: Calculate groundfish discards}
\usage{
discard_groundfish(
  con,
  species = species,
  FY = fy,
  gf_dat = gf_dat,
  non_gf_dat = non_gf_dat,
  gf_trips_only = F,
  save_dir = file.path(getOption("maps.discardsPath"), "groundfish")
)
}
\arguments{
\item{con}{ROracle connection to Oracle (e.g. MAPS)}

\item{species}{dataframe with species info}

\item{FY}{Fishing Year}

\item{gf_dat}{Data frame of groundfish trips built from CAMS_OBS_CATCH and control script routine}

\item{non_gf_dat}{Data frame of non-groundfish trips built from CAMS_OBS_CATCH and control script routine}

\item{gf_trips_only}{Flag if user only wants to run groundfish trips (e.g. for quota monitoring)}

\item{save_dir}{Directory to save (and load saved) results}
}
\value{
nothing currently, writes out to fst files (add oracle?)
}
\description{
discard_groundfish: Calculate groundfish discards
}
