% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discard_groundfish_diagnostic.R
\name{discard_groundfish_diagnostic}
\alias{discard_groundfish_diagnostic}
\title{discard_groundfish_diagnostic: Calculate groundfish discards: Calculate discards for May groundfish fishing year}
\usage{
discard_groundfish_diagnostic(
  con = con_maps,
  species = species,
  FY = fy,
  gf_dat = gf_dat,
  non_gf_dat = non_gf_dat,
  return_table = T,
  return_summary = F,
  CAMS_GEAR_STRATA = CAMS_GEAR_STRATA,
  STOCK_AREAS = STOCK_AREAS,
  CAMS_DISCARD_MORTALITY_STOCK = CAMS_DISCARD_MORTALITY_STOCK
)
}
\arguments{
\item{con}{Oracle connection}

\item{species}{data frame of species for evaluation. Can be a dataframe of multiple species or single row (preferred for testing)}

\item{FY}{fishing year for evaluation}

\item{gf_dat}{Data frame of groundfish trips built from CAMS_OBS_CATCH and control script routine. sourced from CAMS_OBS_CATCH}

\item{non_gf_dat}{Data frame of non-groundfish trips built from CAMS_OBS_CATCH and control script routine. sourced from CAMS_OBS_CATCH}

\item{return_table}{logical, should a table (very large!) of trip level info be returned?}

\item{return_summary}{logical, should a summary (very not large) be returned?}

\item{CAMS_GEAR_STRATA}{support table sourced from Oracle}

\item{STOCK_AREAS}{support table sourced from Oracle}

\item{CAMS_DISCARD_MORTALITY_STOCK}{support table sourced from Oracle}
}
\value{
nothing currently, writes out to fst files (add oracle?)
}
\description{
Support tables in this version ar function variables. This allows them to be modified and run for diagnostic testing. This function can output a summary of discard information by strata, including number of trips used, variance, CV, discard amount, and KALL. It can also output a table, by subtrip, with all information. This table can be summarized afterward as well. If testing several scenarios, it is recommended to only output the summary rather than the full subtrip table as each output is 1-2GB.
}
\details{
The original function pushed .fst files directly to the Wind server at GARFO. This version does not do that and only produces local results.
When running windowpane or yellowtail flounders, this function will save .fst files of scallop trips locally. These will be auto-saved in your working directory.
}
\examples{
see vignette


}
\author{
Benjamin Galuardi
}
