% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discard_herring.R
\name{discard_herring}
\alias{discard_herring}
\title{discard_herring: Calculate discards for January fishing year for herring only}
\usage{
discard_herring(
  con,
  species = species,
  FY = fy,
  all_dat = all_dat,
  save_dir = file.path(getOption("maps.discardsPath"), "herring")
)
}
\arguments{
\item{con}{ROracle connection to Oracle (e.g. MAPS)}

\item{species}{dataframe with species info}

\item{FY}{Fishing Year}

\item{all_dat}{Data frame of trips built from CAMS_OBS_CATCH and control script routine}

\item{save_dir}{Directory to save (and load saved) results}
}
\value{
nothing currently, writes out to fst files (add oracle?)
}
\description{
discard_herring: Calculate discards for January fishing year for herring only
}
