% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discard_herring_diagnostic.R
\name{discard_herring_diagnostic}
\alias{discard_herring_diagnostic}
\title{discard_herring: Calculate discards for January fishing year for herring only}
\usage{
discard_herring_diagnostic(
  con,
  species = species,
  FY = fy,
  herr_dat = herr_dat,
  CAMS_GEAR_STRATA = CAMS_GEAR_STRATA,
  STOCK_AREAS = STOCK_AREAS,
  CAMS_DISCARD_MORTALITY_STOCK = CAMS_DISCARD_MORTALITY_STOCK,
  return_table = T,
  return_summary = F
)
}
\arguments{
\item{con}{ROracle connection to Oracle (e.g. MAPS)}

\item{species}{dataframe with species info}

\item{FY}{Fishing Year}

\item{herr_dat}{Data frame of trips built from CAMS_OBS_CATCH and control script routine}

\item{save_dir}{Directory to save (and load saved) results}
}
\value{
nothing currently, writes out to fst files (add oracle?)
}
\description{
discard_herring: Calculate discards for January fishing year for herring only
}
