% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discard_may.R
\name{discard_may}
\alias{discard_may}
\title{discard_may: Calculate discards for May fishing year species}
\usage{
discard_may(
  con,
  species = species,
  FY = fy,
  all_dat = all_dat,
  save_dir = file.path(getOption("maps.discardsPath"), "may")
)
}
\arguments{
\item{con}{ROracle connection to Oracle (e.g. MAPS)}

\item{species}{dataframe with species info}

\item{FY}{Fishing Year}

\item{all_dat}{Data frame of trips built from CAMS_OBS_CATCH and control script routine}

\item{save_dir}{Directory to save (and load saved) results}
}
\value{
nothing currently, writes out to fst files (add oracle?)
}
\description{
discard_may: Calculate discards for May fishing year species
}
