% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eflalo.r
\docType{data}
\name{eflalo}
\alias{eflalo}
\title{eflalo dataset from VMSTOOLS}
\format{
a dataframe
}
\usage{
data(eflalo)
}
\description{
Fisheries data from ICES, masked for time and individual.
}
\examples{
data(eflalo)
}
\references{
Niels Hintzen, Francois Bastardie and Doug Beare (2014). vmstools: For analysing fisheries VMS (Vessel Monitoring System) data. R package version 0.72. http://CRAN.R-project.org/package=vmstools
}
\seealso{
\code{\link{vmstools}}, \code{\link{make.obs.flag.dat}}
}
\keyword{datasets}
