% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.cochran.ss.by.strat.r
\name{get.cochran.ss.by.strat}
\alias{get.cochran.ss.by.strat}
\title{Estimate discard rates (and other attributes) by strata}
\usage{
get.cochran.ss.by.strat(
  bydat,
  trips,
  targCV = 0.3,
  strata_name = "STRATA",
  strata_complete = c("dredge", "trawl")
)
}
\arguments{
\item{bydat}{Observed trips}

\item{trips}{Total trips (e.g., DMIS data from GARFO)}

\item{targCV}{Target CV for sample size determination (default = 0.3)}

\item{strata_name}{The name of the strata column (defaul = 'STRATA').}

\item{strata_complete}{NOT OPTIONAL! A vector of all unique strata :)
desired from the output (including unobserved).}
}
\value{
a list of Cochran ratio estimaes by strata, Discards, Total Discard in the fishery, Total Cv for the fishery, Required Seadays by strata
}
\description{
For each stratum, returns the Cochran ratio estimate of discard rate along with its CV and sample size requirement for a given target CV.  Also returns other
stratum-specific attributes including N, K, n, and k for a set of total trips and observed trips.
}
\details{
Each data frame input MUST have a stratification column that matches. This may be numeric, logical, factor or character
}
\examples{
get.cochran.ss.by.strat
}
