% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.trans.rate.r
\name{get.trans.rate}
\alias{get.trans.rate}
\title{TRANSITION RATE}
\usage{
get.trans.rate(l_observed_trips, l_assumed_rate, l_inseason_rate, ntrans = 5)
}
\arguments{
\item{l_observed_trips}{number of observed trips in current year}

\item{l_assumed_rate}{discard rate from previous time period}

\item{l_inseason_rate}{current in-season discard rate}

\item{ntrans}{number of trips in current year to use in transitional weighting}
}
\value{
discard rate adjusted using prior years rate
}
\description{
TRANSITION RATE
}
\details{
usually not called directly
}
\examples{

plot(get.trans.rate(l_observed_trips = 1:100, l_assumed_rate = 0.001, l_inseason_rate = .0005, ntrans = 50), typ = 'l', ylab = 'Discard rate', xlab = 'Number of in season trips')

}
\seealso{
\code{\link{cochran.trans.calc}}
}

