% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.weighted.cv.r
\name{get.weighted.cv}
\alias{get.weighted.cv}
\title{Required Samples across a CV range}
\usage{
get.weighted.cv(bydat, dest2)
}
\arguments{
\item{bydat}{bycatch data, formatted for Cochran calculations}

\item{dest2}{results from \code{\link{get.cochran.ss.by.strat}}}
}
\value{
a data frame with required samples
}
\description{
Estimate number of required samples (trips), using weighted mean, across a range of CV values for discard rates
}
\examples{
get.weighted.cv
}
\seealso{
\code{\link{get.bydat}}, \code{\link{cochran_calc_ss}}
}
