% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_catch_obs_diag_nefsc.R
\name{get_catch_obs_diag_nefsc}
\alias{get_catch_obs_diag_nefsc}
\title{Get Catch OBS data at NEFSC}
\usage{
get_catch_obs_diag_nefsc(
  con = con_maps,
  start_year = 2017,
  end_year = 2022,
  mrem_table = mrem_table
)
}
\arguments{
\item{con}{database connection}

\item{start_year}{start year (calendar year)}

\item{end_year}{end year (calendar year)}

\item{mrem_table}{table of MREM KALL adjustment. If workgng from NEFSC, this must be created by the NEFSC user}
}
\value{
a list with the elements needed for discard estimation:
* gf_dat : a data frame of groundfish only trips. This is used only in \link{discard_groundfish.R}
* non_gf_dat : data frame of only non groundfish trips
* all_dat : data frame of all trips
}
\description{
Get Catch OBS data at NEFSC
}
\examples{
# Main example
dat = get_catch_obs(con_maps, 2021, 2022, )
gf_dat = dat$gf_dat
non_gf_dat = dat$non_gf_dat
all_dat = dat$all_dat
rm(dat)
gc()

# Herring example
dat = get_catch_obs_herring(con_maps, 2021, 2022)
gf_dat = dat$gf_dat
non_gf_dat = dat$non_gf_dat
all_dat = dat$all_dat
rm(dat)
gc()

}
