% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_covrow_bg.R
\name{get_covrow_bg}
\alias{get_covrow_bg}
\title{Alt version get_covrow.R}
\usage{
get_covrow_bg(joined_table)
}
\arguments{
\item{joined_table}{}
}
\value{
data frame with additional covariance variables
}
\description{
Alt version get_covrow.R
}
