% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discard_generic.R
\name{get_date_range}
\alias{get_date_range}
\title{Get Date Range
get date range for particular type of fishing year}
\usage{
get_date_range(FY, FY_TYPE)
}
\arguments{
\item{FY}{Fishing Year (e.g. 2020)}

\item{FY_TYPE}{Type of fishing year (e.g. CALENDAR). This is case sensitive as it calls `CFG_DISCARD_RUNID`}
}
\value{
a start and end date
}
\description{
Get Date Range
get date range for particular type of fishing year
}
