% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_obs_disc_values.R
\name{get_obs_disc_vals}
\alias{get_obs_disc_vals}
\title{Get Observed Discards}
\usage{
get_obs_disc_vals(c_o_tab = c_o_dat2, species_itis = "164712")
}
\arguments{
\item{c_o_tab}{table of matched observer and commerical trips}

\item{species_itis}{species of interest using SPECIES_ITIS code}
}
\value{
a tibble with: YEAR, VTRSERNO, GEARTYPE, MESHGROUP,KALL, DISCARD, dk
The stratification variables don't matter so much as the assignment of discard is done using VTR Serial Number.
}
\description{
This function is used to get observed discard values on observed trips. These values are used in place of estimated values for those trips that were observed. This is done at the the sub-trip level. 
This function does not need stratification. Only VTR serial number and an observed discard for desired species
This function utilizes the DISCARD_PRORATE field in CAMS_OBS_CATCH. This value must be used for assigning observed discrd to trips. It is NOT used for d/k calculations since it is pro-rated by unobserved KALL.
}
