% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_in.R
\name{\%!in\%}
\alias{\%!in\%}
\title{notin
the most useful function not included in base R!}
\usage{
x \%!in\% y
}
\arguments{
\item{x:}{vector to check}

\item{y:}{vector to serve as reference}
}
\value{
a vector where x is 'not in' y
}
\description{
notin
the most useful function not included in base R!
}
\examples{
\dontrun{
x <- c()
x \%!in\% y
}
}
