% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_assumed_rate.R
\name{make_assumed_rate}
\alias{make_assumed_rate}
\title{Make an assumed rate}
\usage{
make_assumed_rate(
  bdat,
  species_itis = "164712",
  stratvars = c("GEARTYPE", "meshgroup")
)
}
\arguments{
\item{bdat}{table of observed trips that can include (and should include) multiple years}

\item{species_itis}{species of interest using SPECIES_ITIS code}

\item{stratvars}{Stratification variables. These must be columns available in `bdat`. Not case sensitive. Thiws should be a subset of those used in `make_bdat_focal`.}

\item{year}{Year where discard estimate is needed}
}
\value{
a tibbleS with STRATA, KALL (Kept All), BYCATCH (discard of species), dk (discard rate)
}
\description{
This function is meant to return fallback rates that are more general than a more specific stratification used in `make_bdat_focal`. The discard rates are generated directly to be subsequently substituted where needed.
}
