% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bdat_focal.R
\name{make_bdat_focal}
\alias{make_bdat_focal}
\title{Get Observed trips for discard year
This function utilizes the DISCARD field in CAMS_OBS_CATCH. This value must be used for d/k calculations since it represents the observed part of a trip.}
\usage{
make_bdat_focal(
  bdat,
  species_itis = "164712",
  stratvars = c("GEARTYPE", "meshgroup", "region", "halfofyear")
)
}
\arguments{
\item{bdat}{table of observed trips that can include (and should include) multiple years}

\item{species_itis}{species of interest using SPECIES ITIS code}

\item{stratvars}{Stratification variables. These must be columns available in `bdat`. Not case sensitive.}

\item{year}{Year where discard estimate is needed}
}
\value{
a tibble with LINK1, CAMSID, SUBTRIP, STRATA, KALL, BYCATCH. Kept all (KALL) is rolled up by subtrip. BYCATCH is the observed discard of the species of interest.

This table is used in `discaRd`

the source table (bdat) is created outside of this function in SQL. It can be quite large so it is not done functionally here. See vignette (when it's available..)
}
\description{
Get Observed trips for discard year
This function utilizes the DISCARD field in CAMS_OBS_CATCH. This value must be used for d/k calculations since it represents the observed part of a trip.
}
