% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cams_obdbs_all_years.R
\name{make_cams_obdbs_all_years}
\alias{make_cams_obdbs_all_years}
\title{Make CAMS OBDBS All YEARS
builds table for one calendar year of observer data.}
\usage{
make_cams_obdbs_all_years(con)
}
\arguments{
\item{con}{ROracle connection}
}
\value{
builds a table on Oracle schema specified (e.g. MAPS)
}
\description{
Make CAMS OBDBS All YEARS
builds table for one calendar year of observer data.
}
