% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cams_obs_catch.R
\name{make_cams_obs_catch}
\alias{make_cams_obs_catch}
\title{Make CAMS_OBS_CATCH}
\usage{
make_cams_obs_catch(
  con,
  sql_file = "~/PROJECTS/discaRd/CAMS/SQL/make_cams_obs_catch_Aug2022.sql"
)
}
\arguments{
\item{con}{Oracle connection (ROracle preferred)}

\item{sql_file}{Sql file that builds the table. The SQL lives in another folder within the discaRd project}
}
\value{
Oracle table of merged CAMS landings and observer data
}
\description{
Make CAMS_OBS_CATCH
}
\examples{

# unlock keyring

keyring::keyring_unlock("apsd_ma")
con_maps = apsdFuns::roracle_login(key_name = 'apsd_ma', key_service = 'maps')
sqlfile = paste0(path.package('discaRd'),'/inst/SQL/MERGE_CAMS_OBS_CATCH_AUG2022.sql')
make_cams_obs_catch(con_maps, sql_file = sqlfile)

# make indices

# share to CAMS_GARFO
ROracle::dbSendQuery(con_maps, "GRANT ALL ON CAMS_OBS_CATCH TO CAMS_GARFO WITH GRANT OPTION")

# share to NEFSC
con_cams = apsdFuns::roracle_login(key_name = 'apsd_ma', key_service = 'cams_garfo')

Sys.setenv(TZ = "America/New_York")
Sys.setenv(ORA_SDTZ = "America/New_York")
ROracle::dbSendQuery(con_cams, "CREATE OR REPLACE VIEW CAMS_GARFO.CAMS_OBS_CATCH AS SELECT * FROM MAPS.CAMS_OBS_CATCH")
ROracle::dbSendQuery(con_cams, "GRANT SELECT ON CAMS_GARFO.CAMS_OBS_CATCH TO CAMS_GARFO_FOR_NEFSC")

}
