% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mrem_table.R
\name{make_mrem_table}
\alias{make_mrem_table}
\title{Build MREM KALL adjustment table
make a table for yourself since ITD won't let you make a View... ----
Make MREM adjustment table on the fly
This currently lives at GARFO as a VIew. Thus, it is not ported to NEFSC via TTS. NEFSC IT goons do not allow Oracle View creation so the code can't simply be shared in a normal fashion. This function is the work around. It will generate a MREM KALL adjustment table in the users worksapce. This is needed when pulling data for a CAMS discard diagnostic run.}
\usage{
make_mrem_table(con)
}
\arguments{
\item{con}{database connection}
}
\description{
Build MREM KALL adjustment table
make a table for yourself since ITD won't let you make a View... ----
Make MREM adjustment table on the fly
This currently lives at GARFO as a VIew. Thus, it is not ported to NEFSC via TTS. NEFSC IT goons do not allow Oracle View creation so the code can't simply be shared in a normal fashion. This function is the work around. It will generate a MREM KALL adjustment table in the users worksapce. This is needed when pulling data for a CAMS discard diagnostic run.
}
