% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_strata_desc.R
\name{make_strata_desc}
\alias{make_strata_desc}
\title{make_strata_desc}
\usage{
make_strata_desc(x, remove = FALSE)
}
\arguments{
\item{x}{data.frame with column STRATA_USED and the associated, named columns}

\item{remove}{logical indicating whether to return just summarized descriptions or full data set with the new columns added.}
}
\value{
data frame of trips and discards
}
\description{
make_strata_desc
}
\details{
Make the strata description based on the STRATA_USED and associated columns
}
\examples{

\dontrun{
library(dplyr)
library(tidyr)
library(stringr)

foo <- data.frame(
  STRATA_USED = c("A", "A;B", "B;C", "A;B;C", "A;B", "A", "", NA_character_, NA_character_),
  A = c("A1", "A2", "A3", "A4", "", NA_character_, NA_character_, NA_character_, "A8"),
  B = c("B1", "B2", "B3", "B4", "B5", "B6", NA_character_, NA_character_, NA_character_#' ),
#'   C = c("C1", "C2", "C3", "C4", "C5", "C6", NA_character_, NA_character_, "C8")
)

make_strata_desc(foo)
}
}
\author{
Daniel Hocking, Benjamin Galuardi
}
