% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_discard_diag.R
\name{parse_discard_diag}
\alias{parse_discard_diag}
\title{Parse Discard Table for Diagnostics
Parses discard results, cleans the table (e.g. infinite).}
\usage{
parse_discard_diag(joined_table, FY_species = 2018, date_run = NULL)
}
\arguments{
\item{joined_table}{data framesourced from diagnostic discard run. This table is an intermediary table that must be processed before uploadeing to `CAMS_DISCARD_ALL_YEARS`}

\item{date_run}{either NULL or a specific date object. If `NULL`, it defaults to the date the function was run}

\item{FY}{Fishing Year run}

\item{gf_only}{logical}
}
\value{
a data frame with columns corresponding to `CAMS_DISCARD_ALL_YEARS`
}
\description{
Parse Discard Table for Diagnostics
Parses discard results, cleans the table (e.g. infinite).
}
