% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_upload_discard.R
\name{parse_upload_discard}
\alias{parse_upload_discard}
\title{Parse and Upload Discard Table
Parses discard results, cleans the table (e.g. infinite) and uploads to Oracle. This is designed to be used for a series of single species (itis), for a single fishing year, that reside in a folder. This function will upload all species within the folder that correspond to the provided fishing year.}
\usage{
parse_upload_discard(
  con = con_maps,
  filepath = getOption("maps.discardsPath"),
  FY = 2018,
  gf_only = F,
  date_run = NULL
)
}
\arguments{
\item{con}{ROracle connection to Oracle (e.g. MAPS)}

\item{filepath}{path to .fst discard results}

\item{FY}{Fishing Year to upload, should correspond to those results in filepath}
}
\value{
nothing; cleans, deletes, and appends the Oracle table
}
\description{
N.B.! This function will remove the previous table(s) on MAPS.
}
\examples{
\dontrun{
parse_upload_discard(con = con_maps, filepath = getOption("maps.discardsPath"), FY = 2018)
}
}
