% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discard.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{all_vars}
\alias{any_vars}
\alias{vars}
\alias{starts_with}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{all_vars}}, \code{\link[dplyr]{any_vars}}, \code{\link[dplyr]{vars}}}

  \item{tidyselect}{\code{\link[tidyselect]{starts_with}}}
}}

