% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_discard.R
\name{run_discard}
\alias{run_discard}
\title{Run discaRd end to end}
\usage{
run_discard(
  bdat,
  ddat_focal,
  c_o_tab = c_o_dat2,
  species_itis = "164712",
  stratvars = c("GEARTYPE", "meshgroup", "region", "halfofyear"),
  aidx = c(1, 2)
)
}
\arguments{
\item{bdat}{table of observed trips that can include (and should include) multiple years}

\item{ddat_focal}{table of observed trips for discard year}

\item{c_o_tab}{matched table of observed and commerical trips}

\item{species_itis}{species of interest using SPECIES_ITIS code}

\item{stratvars}{stratification variables desired}

\item{aidx}{subset of `stratvars` for a simplified stratification.}

\item{year}{year where discard is needed}
}
\value{
a list of: 
1) Species, 
2) discaRd results (summary table, CV, etc), 
3)Complete table of commercial trips and discard amounts
}
\description{
Run discaRd end to end
}
