% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scallop_subroutine_diag.R
\name{scallop_subroutine_diag}
\alias{scallop_subroutine_diag}
\title{Groundfish discards on scallop trips
This function oeprates identically to otehr discard modules but is only relevant for scallop trips, and is predicated on what groundfish fishing year is under evaluation.}
\usage{
scallop_subroutine_diag(
  FY = 2019,
  con = con_maps,
  scal_gf_species,
  non_gf_dat = non_gf_dat,
  scal_trip_dir = file.path(getOption("maps.discardsPath"), "scallop_groundfish")
)
}
\arguments{
\item{FY}{Groundfish Fishing year under evaluation. This value determines which scallop fishing years are run}

\item{con}{ROracle database connection}

\item{scal_gf_species}{data frame, single row, that has species_itis and common name of groundfish under evaluation (e.g. yellowtail and windowpane flounder)}

\item{non_gf_dat}{data frame containing non-groundfish trips from `CAMS_LANDINGS`. This is loaded from Oracle from a control script.}

\item{scal_trip_dir}{Directory to save scallop trip results}
}
\value{
writes a .fst file for disacrd estimates on scallop trips for species under evaluation. This folder and will be referenced within the groundfish module.
}
\description{
Groundfish discards on scallop trips
This function oeprates identically to otehr discard modules but is only relevant for scallop trips, and is predicated on what groundfish fishing year is under evaluation.
}
\author{
Ben Galuardi
}
