% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.parallel.r
\name{setup.parallel}
\alias{setup.parallel}
\title{Setup parallel}
\usage{
setup.parallel(n.cores = NULL)
}
\value{
creates settings to allow parallel processing
}
\description{
Setup parallel
}
\examples{
setup.parallel
}
