% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_parallel.r
\name{setup_parallel}
\alias{setup_parallel}
\title{Setup parallel}
\usage{
setup_parallel(n.cores = NULL)
}
\value{
creates settings to allow parallel processing
}
\description{
Setup parallel
}
\examples{
setup.parallel
}
