% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_single_discard_row.R
\name{summarise_single_discard_row}
\alias{summarise_single_discard_row}
\title{Reduce the ddat_focal dataframe for inputting into discards to one row per species-subtrip-link1}
\usage{
summarise_single_discard_row(data, itis_tsn)
}
\arguments{
\item{itis_tsn}{focal species itis code}

\item{ddat_focal}{dataframe to summarise}
}
\value{
ddat_focal dataframe
}
\description{
Reduce the ddat_focal dataframe for inputting into discards to one row per species-subtrip-link1
}
\details{
Reduce the ddat_focal dataframe for inputting into discards to one row per species-subtrip-link1
}
\author{
Daniel J. Hocking

Benjamin Galuardi
}
