% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unite_by_source.R
\name{unite_by_source}
\alias{unite_by_source}
\title{Unite by source
This function gets a `STRATA_USED_DESC` for a data frame. This must be used witha specific `DISCARD_SOURCE` and `GF` grouping.}
\usage{
unite_by_source(mytable)
}
\arguments{
\item{mytable}{input table from discard estiamtion}
}
\value{
table with STRATA_USED_DESC column
}
\description{
Unite by source
This function gets a `STRATA_USED_DESC` for a data frame. This must be used witha specific `DISCARD_SOURCE` and `GF` grouping.
}
\examples{

tab_list =joined_table \%>\%
group_split(DISCARD_SOURCE, GF) \%>\%   # group_split may be deprecated in the future..
  lapply(., unite_by_source)

joined_table = do.call(rbind, tab_list) \%>\%
  as_tibble()

}
\author{
Benjamin Galuardi
}
