% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vary.bigN.r
\name{vary.bigN}
\alias{vary.bigN}
\title{Vary N (# total commercial trips)}
\usage{
vary.bigN(bdat, bigN, by = 10)
}
\arguments{
\item{bdat}{data loaded from \code{\link{get.bdat}}}

\item{bigN}{Total trips}

\item{by}{sequencing step. Start is the number of observed trips, end is big N}
}
\value{
SD and CV from \code{\link{cochran_rse}}
}
\description{
Function to run Cochran estimator with a variable big N (total number of trips). Sequences from the numnber of observed trips up to a maximum number of total trips.  N > n
}
\examples{
vary.bigN
}

