% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_in.R
\name{\%!in\%}
\alias{\%!in\%}
\title{notin}
\usage{
x \%!in\% y
}
\arguments{
\item{x:}{vector to check}

\item{y:}{vector to serve as reference}
}
\description{
notin
}
\examples{
\dontrun{
x <- c()
x \%!in\% y
}
}
